move modules around
