--- lib/typing.py
@@ -1,9 +1,0 @@
-"""Typing annotations."""
-
-import disnake
-
-ApplicationCommandInteraction = disnake.ApplicationCommandInteraction
-User = disnake.User
-Member = disnake.Member
-TextChannel = disnake.TextChannel
-DMChannel = disnake.DMChannel
