+++ slashbot/config.py
